requirements.txt
streamlit
pathlib
pillow
torch 
torchvision
torch-summary
tensorflow-cpu
opencv-python
# nsfw-detector